+++ src/nvim/ex_cmds.c
@@ -3976,7 +3976,7 @@ static int do_sub(exarg_T *eap, proftime_T timeout, long cmdpreview_ns, handle_T
                         _("replace with %s (y/n/a/q/l/^E/^Y)?"), sub);
               msg_no_more = false;
               msg_scroll = (int)i;
-              showruler(true);
+              show_cursor_info(true);
               ui_cursor_goto(msg_row, msg_col);
               RedrawingDisabled = temp;
