+++ src/nvim/statusline.h
@@ -1,0 +1,10 @@
+#ifndef NVIM_STATUSLINE_H
+#define NVIM_STATUSLINE_H
+
+#include "nvim/buffer_defs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "statusline.h.generated.h"
+#endif
+
+#endif  // NVIM_STATUSLINE_H
